feat: allow errors from setCookie to be ignored Refer this: request/request#794
This is an error from tough cookie and they allow us to ignore errors. Sometimes we don't wanna stop the request when there is an error in cookie setting.

So, now we can allow user to ignore or raise error in setCookie. This will be helpful as most of use this package to fetch body from a request and we don't want to end up with an error when there is something wrong in setCookie.