Put type declarations next to JS file (fixes #84)
It seems that setting `"types": "index.d.ts"` in `package.json` doesn't
cut it in all cases, and some build toolchains really want the `.d.ts`
files to be next to the `.js` file it's associated with.

So even if since 3.0.0 the `.d.ts` file is the same for `esm/index.js`
and `cjs/index.js`, we still need to copy an identical declaration file
next to the JS files.